+++ source/view/player/player.hpp
@@ -11,7 +11,7 @@ public:
 	 * @param initial_position - The initial position of the player character.
 	 * @param max_height - The maximum height the player can reach in the game.
 	 */
-	Player(const sf::Vector2f initial_position, const float max_height);
+	Player(const sf::Vector2f initial_position, const double max_height);
 
 	/**
 	 * @brief Destroys the Player object.
@@ -26,7 +26,7 @@ public:
 	 * @param delta_y - The vertical distance by which the player should be
 	 * moved.
 	 */
-	auto move(float delta_y) -> void;
+	auto move(double delta_y) -> void;
 
 	/**
 	 * @brief Returns a reference to the SFML RectangleShape representing the
@@ -40,6 +40,6 @@ public:
 private:
 	sf::Vector2f m_position;
 
-	const float m_max_height;
+	const double m_max_height;
 	sf::RectangleShape m_rect_shape;
 };
